pembuatan sweer alert

+++ app/Livewire/Category/CategoryIndex.php
@@ -4,6 +4,7 @@ namespace App\Livewire\Category;
 
 use Livewire\Component;
 use App\Models\Category;
+use App\Models\Product;
 use Livewire\Attributes\Layout;
 use Livewire\Attributes\On;
 use Livewire\Attributes\Title;
@@ -12,17 +13,30 @@ use Livewire\Attributes\Title;
 class CategoryIndex extends Component
 {
     // #[Layout('template-dashboard.main')]
-    // #[On('modal-close')]
+    // #[On('status')]
     public function render()
     {
         $categories = Category::all();
         return view('livewire.category.category-index', compact('categories'));
     }
 
+    #[On('destroy')] 
     public function destroy($id)
     {
-        Category::find($id)->delete();
+        $product = Product::where('category_id', $id)->count();
 
-        session()->flash('status','Data berhasil dihapus!');
+        if ($product > 0) {
+
+            $this->dispatch('status', text:'Kategori tidak dapat dihapus!', icon:'error', btn:'btn btn-danger');
+
+        } else {
+
+            Category::find($id)->delete();
+
+            $this->dispatch('status', text:'Data berhasil dihapus!', icon:'success',  btn:'btn btn-success');
+        }
+
+
+        // session()->flash('status','Data berhasil dihapus!');
     }
 }
